reformatted code slightly, added whitelist ix (untested)

+++ programs/gem-bank/src/instructions/set_bank_flags.rs
@@ -9,7 +9,7 @@ pub struct SetBankFlags<'info> {
     pub manager: Signer<'info>,
 }
 
-pub fn handler(ctx: Context<SetBankFlags>, flags: u64) -> ProgramResult {
+pub fn handler(ctx: Context<SetBankFlags>, flags: u8) -> ProgramResult {
     let bank = &mut ctx.accounts.bank;
 
     let flags = Bank::read_flags(flags)?;
